Deleting ill-conceived type definitions
These should not exist because they hardwire a static definition of Float to a fixed type. It defeats the purposed of having everything templatized on the <T> so that it works correctly based on context. These definitions would cause confusion if someone uses both float and double in the same app.