Refactor of equality checking
1) Redefine (more correctly) the way that equality checking is done using an epsilon. The previous approach did not correctly scale the epsilon comparison to the magnitude of the operands. This lead to many anomalies such as incorrectly reported degenerate triangles or collinear points when magnitude of the numbers was not close to 1.

2) For the sake of compactness (the more code, the more place for errors) removed the non-epsilon versions of equality checks

3) Made sure that all equality checks pass down the epsilon.

Really, the epsilon should be what it is for the specified precision and the user should not muck with it. In the long run, it would be best to remove the ability entirely. The epsilon() is not an fudge factor but a specific constant related to the floating point precision.